+++ src/tup/server/windepfile.c
@@ -189,20 +189,25 @@ static int create_process(struct server *s, int dfd, char *cmdline,
 	sec.lpSecurityDescriptor = NULL;
 	sec.bInheritHandle = TRUE;
 
-	if(chdir(win32_get_dirpath(tup_top_fd())) < 0) {
-		perror("chdir");
-		fprintf(stderr, "tup error: Unable to chdir to the project root directory to create a temporary output file.\n");
-		return -1;
+	if(!s->streaming_mode) {
+		if(chdir(win32_get_dirpath(tup_top_fd())) < 0) {
+			perror("chdir");
+			fprintf(stderr, "tup error: Unable to chdir to the project root directory to create a temporary output file.\n");
+			return -1;
+		}
+		swprintf(buf, 64, L".tup\\tmp\\output-%i", s->id);
+		buf[63] = 0;
+		sa.hStdOutput = CreateFile(buf, GENERIC_WRITE, 0, &sec, CREATE_ALWAYS, FILE_ATTRIBUTE_TEMPORARY, NULL);
+		if(sa.hStdOutput == INVALID_HANDLE_VALUE) {
+			fprintf(stderr, "tup error: Unable to create temporary file for stdout\n");
+			return -1;
+		}
+		sa.hStdError = sa.hStdOutput;
+	} else {
+		sa.hStdOutput = GetStdHandle(STD_OUTPUT_HANDLE);
+		sa.hStdError = GetStdHandle(STD_ERROR_HANDLE);
 	}
-	swprintf(buf, 64, L".tup\\tmp\\output-%i", s->id);
-	buf[63] = 0;
 	sa.hStdInput = nul_handle;
-	sa.hStdOutput = CreateFile(buf, GENERIC_WRITE, 0, &sec, CREATE_ALWAYS, FILE_ATTRIBUTE_TEMPORARY, NULL);
-	if(sa.hStdOutput == INVALID_HANDLE_VALUE) {
-		fprintf(stderr, "tup error: Unable to create temporary file for stdout\n");
-		return -1;
-	}
-	sa.hStdError = sa.hStdOutput;
 	sa.dwFlags = STARTF_USESTDHANDLES;
 
 	pi->hProcess = INVALID_HANDLE_VALUE;
@@ -238,7 +243,9 @@ static int create_process(struct server *s, int dfd, char *cmdline,
 		NULL,
 		&sa,
 		pi);
-	CloseHandle(sa.hStdOutput);
+	if(!s->streaming_mode) {
+		CloseHandle(sa.hStdOutput);
+	}
 	free(wcmdline);
 
 	if(!ret)
@@ -256,7 +263,6 @@ int server_exec(struct server *s, int dfd, const char *cmd, struct tup_env *newe
 	DWORD return_code = 1;
 	PROCESS_INFORMATION pi;
 	struct estring cmdline;
-	char buf[64];
 	char depfile[PATH_MAX];
 	char vardict_file[PATH_MAX];
 	HANDLE h;
@@ -374,15 +380,18 @@ int server_exec(struct server *s, int dfd, const char *cmd, struct tup_env *newe
 		goto err_terminate;
 	}
 
-	snprintf(buf, sizeof(buf), ".tup/tmp/output-%i", s->id);
-	buf[sizeof(buf)-1] = 0;
-	s->output_fd = openat(tup_top_fd(), buf, O_RDONLY);
-	if(s->output_fd < 0) {
-		pthread_mutex_lock(s->error_mutex);
-		perror(buf);
-		fprintf(stderr, "tup error: Unable to open sub-process output file after the process completed.\n");
-		pthread_mutex_unlock(s->error_mutex);
-		goto end;
+	if(!s->streaming_mode) {
+		char buf[64];
+		snprintf(buf, sizeof(buf), ".tup/tmp/output-%i", s->id);
+		buf[sizeof(buf)-1] = 0;
+		s->output_fd = openat(tup_top_fd(), buf, O_RDONLY);
+		if(s->output_fd < 0) {
+			pthread_mutex_lock(s->error_mutex);
+			perror(buf);
+			fprintf(stderr, "tup error: Unable to open sub-process output file after the process completed.\n");
+			pthread_mutex_unlock(s->error_mutex);
+			goto end;
+		}
 	}
 
 	s->exited = 1;
@@ -419,15 +428,17 @@ err_terminate:
 
 int server_postexec(struct server *s)
 {
-	char buf[64];
-	snprintf(buf, sizeof(buf), ".tup/tmp/output-%i", s->id);
-	buf[sizeof(buf)-1] = 0;
-	if(unlinkat(tup_top_fd(), buf, 0) < 0) {
-		pthread_mutex_lock(s->error_mutex);
-		perror(buf);
-		fprintf(stderr, "tup error: Unable to unlink sub-process output file.\n");
-		pthread_mutex_unlock(s->error_mutex);
-		return -1;
+	if(!s->streaming_mode) {
+		char buf[64];
+		snprintf(buf, sizeof(buf), ".tup/tmp/output-%i", s->id);
+		buf[sizeof(buf)-1] = 0;
+		if(unlinkat(tup_top_fd(), buf, 0) < 0) {
+			pthread_mutex_lock(s->error_mutex);
+			perror(buf);
+			fprintf(stderr, "tup error: Unable to unlink sub-process output file.\n");
+			pthread_mutex_unlock(s->error_mutex);
+			return -1;
+		}
 	}
 	return 0;
 }
